fix cpu build error
