+++ onnxruntime/test/common/tensor_op_test_utils.h
@@ -43,5 +43,15 @@ inline void Normalize(std::vector<float>& v,
                    std::bind(std::divides<float>(), std::placeholders::_1, stdev));
   }
 }
+
+inline std::vector<MLFloat16> ToFloat16(const std::vector<float>& data) {
+  std::vector<MLFloat16> result;
+  result.reserve(data.size());
+  for (size_t i = 0; i < data.size(); i++) {
+    result.push_back(MLFloat16(math::floatToHalf(data[i])));
+  }
+  return result;
+}
+
 }  // namespace test
 }  // namespace onnxruntime
